Added maintenance api, added and fixed documentation. Api return 404 when the object not found

+++ server/controllers/category.controller.ts
@@ -21,6 +21,12 @@ export interface updateCategoryData {
   parentId: any
 }
 
+export interface createInstructionData {
+  title: string,
+  body: string | null,
+  maintenanceId: number
+}
+
 export const getAllCategories = async () => {
   try {
     const allCategories = await prisma.category.findMany({
@@ -42,6 +48,15 @@ export const getAllCategoriesWithDetails = async () => {
         children: true,
         parent: true,
         Device: true,
+        Maintenance: {
+          select: {
+            id: true,
+            name: true,
+            exceptive: true,
+            categoryId: true,
+            Instruction: true
+          }
+        }
       },
     })
     return allCategories
@@ -100,6 +115,13 @@ export const createCategory = async (CategoryData: createCategoryData): Promise<
 
 export const deleteCategoryById = async (id: number) => {
   try {
+    const validCategory = await prisma.category.findFirst({
+      where: {
+        id: id
+      }
+    })
+    if (!validCategory) return { status: 404, message: `Category not found with id: ${id}` }
+
     const deletedCategory = await prisma.category.delete({
       where: {
         id: Number(id)
@@ -116,6 +138,13 @@ export const updateCategoryById = async (id: number, CategoryData: updateCategor
   try {
     let wantToBeCategoryData = CategoryData
 
+    const validCategory = await prisma.category.findFirst({
+      where: {
+        id: id
+      }
+    })
+    if (!validCategory) return { status: 404, message: `Category not found with id: ${id}` }
+
     const updatedCategory = await prisma.category.update({
       where: {
         id: Number(id),
@@ -128,5 +157,17 @@ export const updateCategoryById = async (id: number, CategoryData: updateCategor
   } catch (error: any) {
     throw new Error(error)
 
+  }
+}
+
+//TODO: Ezt átrakni a megfelelő helyre
+export const createInstructionForMaintenanceById = async (id: number, InstructionData: createInstructionData) => {
+  try {
+    const createdInstruction = await prisma.instruction.create({
+      data: { ...InstructionData, maintenanceId: id }
+    })
+    return createdInstruction
+  } catch (error: any) {
+    throw new Error(error)
   }
 }
